+++ packages/atomic-hosted-page/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.2.43
+
+### Patch Changes
+
+- Updated dependencies [[`2abc0a0`](https://github.com/coveo/ui-kit/commit/2abc0a0cf854db560cf56e9337314c3736c45d88)]:
+  - @coveo/headless@3.51.2
+
 ## 1.2.42
 
 ### Patch Changes